5 задание доработка по замечаниям наставника
